Move sixtyfps::ComponentHandle and sixtyfps::Weak into corelib::api
This allows re-use of these types from node and the interpreter without
depending on the public Rust API crate, which is a heavy dependency.
