+++ api/sixtyfps-node/native/Cargo.toml
@@ -24,7 +24,6 @@ once_cell = "1.5"
 sixtyfps-compilerlib = { version = "=0.2.0", path="../../../sixtyfps_compiler" }
 sixtyfps-interpreter = { version = "=0.2.0", path="../../../sixtyfps_runtime/interpreter",  features = ["display-diagnostics"] }
 sixtyfps-corelib = { version = "=0.2.0", path="../../../sixtyfps_runtime/corelib" }
-sixtyfps = { version = "=0.2.0", path="../../../api/sixtyfps-rs", default-features = false, features = ["std"] }
 scoped-tls-hkt = "0.1"
 neon = "0.8.0"
 css-color-parser2 = "1.0.1"
