+++ api/sixtyfps-node/native/lib.rs
@@ -4,11 +4,11 @@
 use core::cell::RefCell;
 use neon::prelude::*;
 use rand::RngCore;
-use sixtyfps::ComponentHandle;
 use sixtyfps_compilerlib::langtype::Type;
 use sixtyfps_corelib::model::{Model, ModelHandle};
 use sixtyfps_corelib::window::WindowHandleAccess;
 use sixtyfps_corelib::{ImageInner, SharedVector};
+use sixtyfps_interpreter::ComponentHandle;
 use std::rc::Rc;
 
 mod js_model;
